refactor(calculator): utils 로직을 utils 패키지 내 클래스로 분리
