+++ src/main/java/calculator/StringCalculator.java
@@ -1,51 +1,21 @@
 package calculator;
 
+import calculator.utils.SplitByDelimiterToNumbersGenerator;
 import calculator.validator.InputValidator;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class StringCalculator {
-    private static final int SEPARATOR_END_SIGNATURE_LENGTH = 2;
-    private static final String CUSTOM_SEPARATOR_START_SIGNATURE = "//";
-    private static final String CUSTOM_SEPARATOR_END_SIGNATURE = "\\n";
-    private static final String JOINING_SEPARATOR_DELIMITER = "|";
-    private static final String BASIC_SEPARATOR_COMMA = ",";
-    private static final String BASIC_SEPARATOR_COLON = ":";
     private static final int NULL_OR_EMPTY_VALUE = 0;
 
     public int calculate(String input) {
         if (InputValidator.isNullOrEmpty(input)) {
             return NULL_OR_EMPTY_VALUE;
         }
-        List<String> inputNumbers = split(input);
+        List<String> inputNumbers = SplitByDelimiterToNumbersGenerator.split(input);
         InputValidator.checkRightPositiveInteger(inputNumbers);
         return inputNumbers.stream()
             .mapToInt(Integer::valueOf)
             .sum();
-    }
-
-    private static List<String> split(String input) {
-        List<String> delimiters = new ArrayList<>(Arrays.asList(BASIC_SEPARATOR_COMMA, BASIC_SEPARATOR_COLON));
-        String parameterString = getParameterString(input, delimiters);
-        String delimiter = String.join(JOINING_SEPARATOR_DELIMITER, delimiters);
-        return Arrays.asList(parameterString.split(delimiter));
-    }
-
-    private static String getParameterString(String input, List<String> delimiters) {
-        if (hasCustomSeparator(input)) {
-            String customDelimiter = input.substring(
-                input.indexOf(CUSTOM_SEPARATOR_START_SIGNATURE) + CUSTOM_SEPARATOR_START_SIGNATURE.length(),
-                input.indexOf(CUSTOM_SEPARATOR_END_SIGNATURE)
-            );
-            delimiters.add(customDelimiter);
-            return input.substring(input.indexOf(CUSTOM_SEPARATOR_END_SIGNATURE) + SEPARATOR_END_SIGNATURE_LENGTH);
-        }
-        return input;
-    }
-
-    private static boolean hasCustomSeparator(String input) {
-        return input.contains(CUSTOM_SEPARATOR_START_SIGNATURE) && input.contains(CUSTOM_SEPARATOR_END_SIGNATURE);
     }
 }
